requirements.txt
fastapi==0.135.1
uvicorn==0.42.0
psycopg2-binary==2.9.11
python-dotenv==1.2.2
pydantic==2.11.3
python-jose[cryptography]==3.5.0
